+++ optaweb-vehicle-routing-backend/src/main/java/org/optaweb/vehiclerouting/service/distance/DistanceMatrixImpl.java
@@ -30,7 +30,7 @@ class DistanceMatrixImpl implements DistanceMatrix {
 
     private final DistanceCalculator distanceCalculator;
     private final DistanceRepository distanceRepository;
-    private final Map<Location, Map<Long, Double>> matrix = new HashMap<>();
+    private final Map<Location, Map<Long, Long>> matrix = new HashMap<>();
 
     @Autowired
     DistanceMatrixImpl(DistanceCalculator distanceCalculator, DistanceRepository distanceRepository) {
@@ -48,17 +48,17 @@ class DistanceMatrixImpl implements DistanceMatrix {
         // The map must be thread-safe because:
         // - we're updating it from the parallel stream below
         // - it is accessed from solver thread!
-        Map<Long, Double> distancesToOthers = new ConcurrentHashMap<>(); // the new row
+        Map<Long, Long> distancesToOthers = new ConcurrentHashMap<>(); // the new row
 
         // distance to self is 0
-        distancesToOthers.put(newLocation.id(), 0.0);
+        distancesToOthers.put(newLocation.id(), 0L);
 
         // for all entries (rows) in the matrix:
         matrix.entrySet().stream().parallel().forEach(distanceRow -> {
             // entry key is the existing (other) location
             Location other = distanceRow.getKey();
             // entry value is the data (cells) in the row (distances from the entry key location to any other)
-            Map<Long, Double> distancesFromOther = distanceRow.getValue();
+            Map<Long, Long> distancesFromOther = distanceRow.getValue();
             // add a new cell to the row with the distance from the entry key location to the new location
             // (results in a new column at the end of the loop)
             distancesFromOther.put(newLocation.id(), calculateOrRestoreDistance(other, newLocation));
@@ -69,8 +69,8 @@ class DistanceMatrixImpl implements DistanceMatrix {
         matrix.put(newLocation, distancesToOthers);
     }
 
-    private double calculateOrRestoreDistance(Location from, Location to) {
-        double distance = distanceRepository.getDistance(from, to);
+    private Long calculateOrRestoreDistance(Location from, Location to) {
+        long distance = distanceRepository.getDistance(from, to);
         if (distance < 0) {
             distance = distanceCalculator.travelTimeMillis(from.coordinates(), to.coordinates());
             distanceRepository.saveDistance(from, to, distance);
@@ -79,7 +79,7 @@ class DistanceMatrixImpl implements DistanceMatrix {
     }
 
     @Override
-    public Map<Long, Double> getRow(Location location) {
+    public Map<Long, Long> getRow(Location location) {
         return matrix.get(location);
     }
